requirements.txt
Flask==1.1.2
PyYAML==5.4.1
psycopg2-binary==2.9.1
gunicorn==20.1.0
requests==2.26.0
